Split file and refactor a bit
