+++ packages/react/src/KeybindingHint/KeybindingHint.tsx
@@ -3,7 +3,7 @@ import type {ReactNode} from 'react'
 import React, {Fragment, memo} from 'react'
 import VisuallyHidden from '../_VisuallyHidden'
 
-import {isMacOS} from '@primer/behaviors/utils'
+import {accessibleKeyName, condensedKeyName, fullKeyName} from './key-names'
 
 export interface KeybindingHintProps {
   /**
@@ -35,113 +35,6 @@ export interface KeybindingHintProps {
 type KeybindingHintFormat = 'condensed' | 'full'
 
 type KeybindingHintVariant = 'normal' | 'onEmphasis'
-
-// In the below records, we don't intend to cover every single possible key - only those that
-// would be realistically used in shortcuts. For example, the Pause/Break key is not necessary
-// because it is not found on many keyboards.
-
-/**
- * Short-form iconic versions of keys. These should be intuitive and match icons on keyboards.
- */
-const condensedKeys = (): Record<string, string> => ({
-  alt: isMacOS() ? '⌥' : 'Alt', // the alt key _is_ the option key on MacOS - in the browser there is no "option" key
-  control: '⌃',
-  shift: '⇧',
-  meta: isMacOS() ? '⌘' : 'Win',
-  mod: isMacOS() ? '⌘' : '⌃',
-  pageup: 'PgUp',
-  pagedown: 'PgDn',
-  arrowup: '↑',
-  arrowdown: '↓',
-  arrowleft: '←',
-  arrowright: '→',
-  plus: '+', // needed to allow +-separated key names
-  backspace: '⌫',
-  delete: 'Del',
-  space: '␣', // allow consumers to use the word "Space" even though it's not the browser key name, because it's more readable in props
-  tab: '⇥',
-  enter: '⏎',
-  escape: 'Esc',
-  function: 'Fn',
-  capslock: 'CapsLock',
-  insert: 'Ins',
-  printscreen: 'PrtScn',
-})
-
-/**
- * Specific key displays for 'full' format. We still do show some icons (ie punctuation)
- * because that's more intuitive, but for the rest of keys we show the standard key name.
- */
-const fullKeys = (): Record<string, string> => ({
-  alt: isMacOS() ? 'Option' : 'Alt',
-  mod: isMacOS() ? 'Command' : 'Control',
-  '+': 'Plus',
-  pageup: 'Page Up',
-  pagedown: 'Page Down',
-  arrowup: 'Up Arrow',
-  arrowdown: 'Down Arrow',
-  arrowleft: 'Left Arrow',
-  arrowright: 'Right Arrow',
-  capslock: 'Caps Lock',
-  printscreen: 'Print Screen',
-})
-
-/**
- * Accessible key names intended to be read by a screen reader. This prevents screen
- * readers from expressing punctuation in speech, ie, reading a long pause instead of the
- * word "period".
- */
-const keyDescriptions = (): Record<string, string> => ({
-  alt: isMacOS() ? 'option' : 'alt',
-  meta: isMacOS() ? 'command' : 'Windows',
-  mod: isMacOS() ? 'command' : 'control',
-  // Screen readers may not be able to pronounce concatenated words - this provides a better experience
-  pageup: 'page up',
-  pagedown: 'page down',
-  arrowup: 'up arrow',
-  arrowdown: 'down arrow',
-  arrowleft: 'left arrow',
-  arrowright: 'right arrow',
-  capslock: 'caps lock',
-  printscreen: 'print screen',
-  // We don't need to represent _every_ symbol - only those found on standard keyboards.
-  // Other symbols should be avoided as keyboard shortcuts anyway.
-  // These should match the colloqiual names of the keys, not the names of the symbols. Ie,
-  // "Equals" not "Equal Sign", "Dash" not "Minus", "Period" not "Dot", etc.
-  '`': 'backtick',
-  '~': 'tilde',
-  '!': 'exclamation point',
-  '@': 'at',
-  '#': 'hash',
-  $: 'dollar sign',
-  '%': 'percent',
-  '^': 'caret',
-  '&': 'ampersand',
-  '*': 'asterisk',
-  '(': 'left parenthesis',
-  ')': 'right parenthesis',
-  _: 'underscore',
-  '-': 'dash',
-  '+': 'plus',
-  '=': 'equals',
-  '[': 'left bracket',
-  '{': 'left curly brace',
-  ']': 'right bracket',
-  '}': 'right curly brace',
-  '\\': 'backslash',
-  '|': 'pipe',
-  ';': 'semicolon',
-  ':': 'colon',
-  "'": 'single quote',
-  '"': 'double quote',
-  ',': 'comma',
-  '<': 'left angle bracket',
-  '.': 'period',
-  '>': 'right angle bracket',
-  '/': 'forward slash',
-  '?': 'question mark',
-  ' ': 'space',
-})
 
 /**
  * Consistent sort order for modifier keys. There should never be more than one non-modifier
@@ -197,27 +90,10 @@ interface KeyProps {
   format: KeybindingHintFormat
 }
 
-/**
- * Converts the first character of the string to upper case and the remaining to lower case.
- */
-// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-const capitalize = ([first, ...rest]: string) => (first?.toUpperCase() ?? '') + rest.join('').toLowerCase()
-
-const keyToAccessibleString = (name: string) => keyDescriptions()[name] || name
-
 const Key = ({name, format}: KeyProps) => (
-  // We represent each individual key as a <kbd> inside a single container <kbd> element.
-  // This requires a bit more styling to override the defaults but is the most semantic way
-  // to do it:
-  //
-  //  > To describe an input comprised of multiple keystrokes, you can nest multiple <kbd>
-  //  > elements, with an outer <kbd> element representing the overall input and each
-  //  > individual keystroke or component of the input enclosed within its own <kbd>.
-  //  > (https://developer.mozilla.org/en-US/docs/Web/HTML/Element/kbd#representing_keystrokes_within_an_input)
   <>
-    <VisuallyHidden>{keyToAccessibleString(name)}</VisuallyHidden>
-    {/* eslint-disable-next-line @typescript-eslint/no-unnecessary-condition */}
-    <span aria-hidden>{(format === 'condensed' ? condensedKeys()[name] : fullKeys()[name]) ?? capitalize(name)}</span>
+    <VisuallyHidden>{accessibleKeyName(name)}</VisuallyHidden>
+    <span aria-hidden>{format === 'condensed' ? condensedKeyName(name) : fullKeyName(name)}</span>
   </>
 )
 
@@ -230,7 +106,7 @@ const splitChord = (chord: string) =>
 
 const compareLowercaseKeys = (a: string, b: string) => getKeySortPriorityValue(a) - getKeySortPriorityValue(b)
 
-const chordToAccessibleString = (chord: string) => splitChord(chord).map(keyToAccessibleString).join(' ')
+const accessibleChordString = (chord: string) => splitChord(chord).map(accessibleKeyName).join(' ')
 
 const Chord = ({keys, format = 'condensed', variant = 'normal'}: KeybindingHintProps) => (
   <Text
@@ -268,8 +144,8 @@ const Chord = ({keys, format = 'condensed', variant = 'normal'}: KeybindingHintP
 
 const splitSequence = (sequence: string) => sequence.split(' ')
 
-const sequenceToAccessibleString = (sequence: string) =>
-  splitSequence(sequence).map(chordToAccessibleString).join(', then ')
+const accessibleSequenceString = (sequence: string) =>
+  splitSequence(sequence).map(accessibleChordString).join(', then ')
 
 /**
  * Indicates the presence of a keybinding available for an action.
@@ -303,4 +179,4 @@ KeybindingHint.displayName = 'KeybindingHint'
  * NOTE that this string should _only_ be used when building `aria-label` or `aria-description` props (never rendered
  * visibly) and should nearly always also be paired with a visible hint for sighted users.
  */
-export const getAccessibleKeybindingHintString = (keys: string) => sequenceToAccessibleString(keys)
+export const getAccessibleKeybindingHintString = (keys: string) => accessibleSequenceString(keys)
